# docs/server/openapi.yaml
openapi: 3.1.0
info:
  title: Telegram To-Do Mini App API
  version: 0.3.0
  description:
    "Telegram To-Do Mini App \u7684\u6240\u6709 HTTP API\uFF0C\u4EE5\u7EDF\
    \u4E00\u7684 OpenAPI \u89C4\u8303\u5448\u73B0\uFF0C\u65B9\u4FBF\u524D\u540E\u7AEF\
    \u3001QA\n\u4E0E\u5916\u90E8\u7CFB\u7EDF\u5FEB\u901F\u5BF9\u9F50\u3002\u6240\u6709\
    \u666E\u901A\u4E1A\u52A1\u63A5\u53E3\u9ED8\u8BA4\u91C7\u7528\u7EDF\u4E00\u54CD\
    \u5E94\u5305\u88F9\uFF1A\n\n```json\n{\n  \"success\": true,\n  \"data\": {},\n\
    \  \"error\": null,\n  \"meta\": {}\n}\n```\n\n- \u9274\u6743\uFF1AMini App \u901A\
    \u8FC7 `X-Telegram-Init-Data` \u8BF7\u6C42\u5934\u4F20\u5165 Telegram \u63D0\u4F9B\
    \u7684 init data\uFF0C\u540E\u7AEF\u9A8C\u8BC1\u540E\u83B7\u5F97\u7528\u6237\u8EAB\
    \u4EFD\u3002\n- \u65F6\u95F4\uFF1A\u9664\u975E\u7279\u6B8A\u8BF4\u660E\uFF0C\u5747\
    \u4E3A ISO8601 UTC \u65F6\u95F4\u5B57\u7B26\u4E32\uFF0C\u5BA2\u6237\u7AEF\u53EF\
    \u6309\u65F6\u533A\u6E32\u67D3\u3002\n"
servers:
  - url: https://api.tg-todo.local
    description: "\u672C\u5730/\u5185\u7F51\u8054\u8C03\u73AF\u5883"
  - url: https://api.tg-todo.app
    description: "\u6B63\u5F0F\u751F\u4EA7\u73AF\u5883"
...
components:
  securitySchemes:
    TelegramInitData:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description:
        "Mini App \u521D\u59CB\u5316\u6570\u636E\uFF0C\u540E\u7AEF\u4F1A\
        \u6821\u9A8C\u7B7E\u540D\u5E76\u6620\u5C04 user_id/chat_id\u3002\u6240\u6709\
        \u9700\u8981\u767B\u5F55\u6001\u7684\u63A5\u53E3\u5747\u9700\u643A\u5E26\u3002\
        \n"
    TelegramWebhookSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description:
        "Telegram Webhook \u4E13\u7528\u5BC6\u94A5\uFF0C\u4E0E BotFather\
        \ \u914D\u7F6E\u7684 secret token \u5BF9\u9F50\uFF0C\u7528\u4E8E\u963B\u6B62\
        \u4F2A\u9020\u8BF7\u6C42\u3002\n"
  parameters:
...
    DatabaseSearchParam:
      name: search
      in: query
      schema:
        type: string
        maxLength: 64
      description:
        "\u6570\u636E\u5E93\u7B5B\u9009\u5173\u952E\u5B57\uFF0C\u652F\u6301\
        \u6309\u540D\u79F0/\u5DE5\u4F5C\u533A\u6A21\u7CCA\u5339\u914D\uFF1B\u4E3A\u7A7A\
        \u65F6\u8FD4\u56DE\u6700\u8FD1\u5E38\u7528\u5217\u8868\u3002\n"
...
    RoleParam:
      name: role
      in: query
      required: false
      schema:
        type: string
        enum:
          - admin
          - member
        default: admin
      description:
        "\u7FA4\u7EC4\u89D2\u8272\u8FC7\u6EE4\u6761\u4EF6\uFF0C\u9ED8\u8BA4\
        \u4EC5\u5C55\u793A\u5F53\u524D\u7528\u6237\u4F5C\u4E3A\u7BA1\u7406\u5458\u7684\
        \u7FA4\u7EC4\uFF0C\u4FBF\u4E8E\u96C6\u4E2D\u7BA1\u7406\u7ED1\u5B9A\u5173\u7CFB\
        \u3002\n"
  schemas:
    Envelope:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description:
            "\u8BF7\u6C42\u662F\u5426\u6210\u529F\uFF0C\u5931\u8D25\u65F6\
            \u9700\u8981\u67E5\u770B `error`\u3002"
        data:
          nullable: true
          description:
            "\u4E1A\u52A1\u6570\u636E\u4E3B\u4F53\uFF0C\u7ED3\u6784\u56E0\
            \u63A5\u53E3\u800C\u5F02\u3002"
        error:
          $ref: "#/components/schemas/ErrorObject"
        meta:
          type: object
          nullable: true
          description:
            "\u989D\u5916\u5143\u6570\u636E\uFF08\u5982\u5206\u9875\u4FE1\
            \u606F\uFF09\uFF0C\u4E3A\u4FDD\u6301\u6269\u5C55\u6027\u9ED8\u8BA4\u53EF\
            \u4E3A\u7A7A\u5BF9\u8C61\u3002"
      description:
        "\u540E\u7AEF HTTP API \u7684\u7EDF\u4E00\u54CD\u5E94\u7ED3\u6784\
        \u3002"
    ErrorObject:
      type: object
      properties:
        code:
          type: string
          description:
            "\u673A\u5668\u53EF\u8BFB\u9519\u8BEF\u7801\uFF0C\u4F8B\u5982\
            \ `unauthorized`\u3002"
        message:
          type: string
          description:
            "\u4FBF\u4E8E\u7528\u6237\u7406\u89E3\u7684\u4E2D\u6587/\u82F1\
            \u6587\u63CF\u8FF0\u3002"
        details:
          type: object
          additionalProperties: true
          nullable: true
          description:
            "\u53EF\u9009\u7684\u8C03\u8BD5\u5B57\u6BB5\uFF08\u5B57\u6BB5\
            \u7F3A\u5931\u7B49\uFF09\u3002"
      required:
        - code
        - message
...
    DatabaseSummary:
      type: object
      required:
        - id
        - name
        - workspace
      properties:
        id:
          type: string
          description:
            "Notion \u6570\u636E\u5E93 ID\uFF0C\u7EDF\u4E00\u4E3A slug/UUID\
            \ \u5B57\u7B26\u4E32\u3002"
        name:
          type: string
          description:
            "Notion \u4E2D\u5C55\u793A\u7684\u6570\u636E\u5E93\u540D\u79F0\
            \u3002"
        workspace:
          type: string
          description:
            "\u6240\u5C5E\u5DE5\u4F5C\u533A\u6216\u4E2A\u4EBA\u7A7A\u95F4\
            \uFF0C\u53EF\u5E2E\u52A9\u533A\u5206\u76F8\u4F3C\u6570\u636E\u5E93\u3002"
        icon:
          type: string
          nullable: true
          description:
            "\u63A8\u8350\u4F7F\u7528 RemixIcon \u7B49 icon \u5173\u952E\
            \u5B57\u6216 URL\u3002"
    GroupSummary:
      type: object
      required:
        - id
        - title
        - status
        - role
      properties:
        id:
          type: string
          description:
            "\u7FA4\u7EC4\u5728\u7CFB\u7EDF\u4E2D\u7684\u552F\u4E00 ID\uFF08\
            \u901A\u5E38\u4E0E Telegram chat_id \u7ED1\u5B9A\uFF09\u3002"
        title:
          type: string
          description: "\u7FA4\u7EC4\u6807\u9898\u3002"
        status:
          type: string
          enum:
            - Connected
            - Unbound
            - Inactive
          description:
            "\u7ED1\u5B9A\u72B6\u6001\uFF0CConnected \u8868\u793A\u5DF2\
            \u5173\u8054\u6570\u636E\u5E93\u3002"
        db:
          $ref: "#/components/schemas/DatabaseSummary"
          nullable: true
          description:
            "\u5DF2\u7ED1\u5B9A\u7684 Notion \u6570\u636E\u5E93\uFF0C\u672A\
            \u7ED1\u5B9A\u65F6\u4E3A null\u3002"
        role:
          type: string
          enum:
            - Admin
            - Member
          description:
            "\u5F53\u524D\u7528\u6237\u5728\u7FA4\u7EC4\u4E2D\u7684\u89D2\
            \u8272\u3002"
...
    DatabaseValidationResult:
      type: object
      required:
        - id
        - name
        - compatible
        - required_fields
        - missing_fields
      properties:
        id:
          type: string
        name:
          type: string
        compatible:
          type: boolean
          description:
            "\u662F\u5426\u6EE1\u8DB3\u6240\u6709\u5FC5\u586B\u5B57\u6BB5\
            \u8981\u6C42\u3002"
        required_fields:
          type: array
          items:
            type: string
        missing_fields:
          type: array
          items:
            type: string
          description:
            "\u53EF\u8BFB\u6027\u8BF4\u660E\uFF0C\u4F8B\u5982\u201C\u7ED3\
            \u6784\u5B8C\u7F8E\u5339\u914D\u201D\u3002"
...
paths:
...
  /databases:
    get:
      tags:
        - Tasks
        - Groups
      security:
        - TelegramInitData: []
      summary: "\u67E5\u8BE2\u53EF\u7528\u7684 Notion \u6570\u636E\u5E93"
      operationId: listDatabases
      parameters:
        - $ref: "#/components/parameters/DatabaseSearchParam"
      responses:
        "200":
          description: "\u6570\u636E\u5E93\u5217\u8868"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/DatabaseSummary"
  /databases/{database_id}/validate:
    get:
      tags:
        - Groups
      security:
        - TelegramInitData: []
      summary: "\u6821\u9A8C\u624B\u52A8\u8F93\u5165\u7684\u6570\u636E\u5E93"
      operationId: validateDatabase
      parameters:
        - name: database_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "\u8FD4\u56DE\u5B57\u6BB5\u68C0\u67E5\u7ED3\u679C"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DatabaseValidationResult"
...
  /groups:
    get:
      tags:
        - Groups
      security:
        - TelegramInitData: []
      summary: "\u7BA1\u7406\u7AEF\u7FA4\u7EC4\u5217\u8868"
      operationId: listGroups
      parameters:
        - $ref: "#/components/parameters/RoleParam"
      responses:
        "200":
          description: "\u8FD4\u56DE\u7FA4\u7EC4\u53CA\u7ED1\u5B9A\u72B6\u6001"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/GroupSummary"
...